signal end of frame when need more bytes from source
..not when Forth is done processing a line

This makes the whole process more intuitive and manageable. When going the
internal Forth approach, you need to concider every instance where Forth is
truely done processing, or is still defining a word when reaching
end-of-line. But we only need extra chars when we are truely done with
processing and we are not executing anything that doesn't need more input.

We also convert the Forth sources back to unix format, without the \r line
endings. And we take away the tabs in DEMO that for some reason still to be
investigated seemed to stop compiling the DEMO word when loading interactively,
while still finishing processing the file. This however was no issue when
loading from gbfs..