start of host Forth compiler: Forth constants to assembly
Adding all of these defines a patch ago makes compiling take a lot longer. We
were looking at about 27 seconds, and in the end there wasn't much code written.

But we can compile Forth words without TOO much hassle. Starting with the raft
of new constants is low-hanging fruit, as constants are trivial to parse. I've
written the most basic of compilers. With the constants out of the way, the
compile time when starting up the GBA is down from that 27 to 7 seconds.

In the future we will also parse variables and functions. The latter in various
forms of complexity. I've already wondered how far in the rabbit hole I should
go, up till a full Forth implementation on the host and emulation of the GBA
registers. Let's take the practical route for now.