integrate Apex Audio System music mixing
I've integrated AAS in the C code, and added some hooks to service/bdos to be
able to control the music from Forth. It's early days, and the integration isn't
top notch. Perhaps in the future, I'll write Forth AAS bindings, but this is
fine for now.

It was a tiny bit tricky to figure out what was needed. Turns out I needed to
use the AAS_FastTimer1InterruptHandler as the irq timer 1 interrupt handler. AAS
automatically configures setting the timer from within the AAS lib.

Also I don't think the lib was written with stopping the loop in mind or with
long blocks of the IRQ, which I can do with the serial interface. But you can
stop the song, and then I think you just need to flush the buffers for a few
vblanks with AAS_DoWork() if you don't want to crash the system. It seemed to
work in the end:

Starting and stopping a game loop with music toggling is
working fine-ish. YAY!