In Makefile, check if AAS_Data.* files exist yet
they created as part of the rule chain, and are not necessarily included in the
list of assembly files when they are scanned in the beginning of the
makefile. When first run the files don't exist yet so they need to be added. On
subsequent runs they files exist from a previous run and shouldn't be added
twice.

It's all a horrible hack really. Not happy at all, but should rewrite this all
to begin with, so am allowing temporary ugliness.