convert metadata of Tiled json file to binary file for GBA inclusion
For now this script just converts the `collision' layer to a binary array of
bytes, which is output to binary file. Later different events can be layered
within the bytes for each tile.

This is quite a space-inefficient approach, but we can easily afford it for
now. At some point I'd like to move to perhaps an infinity tile engine, and
we'll need to rethink all of this anyways.