add sprite bounds checking
by adding a map with collision data and points of interest per tile, used to
check collision box of a sprite.

this needed some heavier refactoring including an abstract sprite object that
holds next to a pointer to the shadow oam control bytes, also holds a bounding
box, abstract object coordinates, object actions, object direction

we also introduce absolute coordinates on the map as the base to calculate bg
offset and sprite offsets on the screen.

and we now have an abstract point struct which is just a word cosisting of a
halfword for x and a halfword for y. that will cover any size of map that I will
ever make.