do more proper sprite<->bg bounds checking
before, we just checked the corners of a sprite. which is a nice beginning, but
that means that a sprite could be impaled by bg thinner than a sprite.

The new code does some horrible stack manipulation as more than 3 vars on the
stack become a hassle to juggle. especially when you can't use the return stack