rewrote Forth `move' word in assembly for mucho speedup
I was running into music glitches when transitioning from the title screen to
main game loop, as the game was copying a lot of data from rom into vram. It did
this by using the most efficient Forth word for this: move. Unfortunately move
was written in Forth, which made it not efficient at all, and the music buffers
couldn't be filled in time while copying. Rewriting it in assembly made all my
timing issues go away.