fix sprite movement after dialog issues
Dialogs were triggered without processing the changed game state for that frame
yet. Once the initial dialog was done, the game loop of the original frame
completed, which caused staggered movement after the dialog. The fix is to delay
displaying the dialog to the next frame. This also gives us an opportunity to
put the sprite into the first idle frame before displaying the dialog. Lastly
this fixes a bug where the sprite would stay in it's walking animation after the
dialog sequence ended.